chart 추가 데이터 입력 및 API, 클라이언트 구현

+++ server/migrations/20211201053655-create-pmi-index.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.DATE
       },
       value: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(3,1)
       },
       createdAt: {
         allowNull: false,
